% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_trial.R
\name{analyse_trial}
\alias{analyse_trial}
\title{Analyse Simulated Trial}
\usage{
analyse_trial(df, time_var, status_var)
}
\arguments{
\item{df}{Input trial dataset.}

\item{time_var}{Time variable name.}

\item{status_var}{Status variable name.}
}
\value{
A data.frame containing HR, p-value, median survival, survival rate and censoring rate.
}
\description{
Calls server API to run Cox model and survival summaries.
}
