% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_base_size.R
\name{auto_base_size}
\alias{auto_base_size}
\title{Automatically Determine Base Font Size}
\usage{
auto_base_size(width = 10, height = 6, min_size = 10, max_size = 16)
}
\arguments{
\item{width}{Plot width.}

\item{height}{Plot height.}

\item{min_size}{Minimum font size.}

\item{max_size}{Maximum font size.}
}
\value{
Numeric base font size.
}
\description{
Calls server API to determine base font size.
}
