% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_limits.R
\name{auto_limits}
\alias{auto_limits}
\title{Automatically Determine Axis Limits}
\usage{
auto_limits(x, expand = 0.06, hard_limits = NULL, include_zero = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{expand}{Expansion ratio.}

\item{hard_limits}{Optional hard limits.}

\item{include_zero}{Whether to include zero.}
}
\value{
Numeric vector of length 2.
}
\description{
Calls server API to determine axis limits.
}
