% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api_url}
\alias{get_api_url}
\title{Get API Server URL}
\usage{
get_api_url()
}
\value{
Character string. The current API server URL.
}
\description{
Get the current base URL of the OncoSeqOS API server.
}
