% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_job_status}
\alias{get_job_status}
\title{Get Job Status}
\usage{
get_job_status(job_id)
}
\arguments{
\item{job_id}{Character string. Job ID returned by `submit_grid_simulation()`
or `run_grid_simulation(wait = FALSE)`.}
}
\value{
A list containing job status information.
}
\description{
Query the current status of an asynchronous simulation job.
}
