% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lambda_from_median}
\alias{lambda_from_median}
\title{Hazard rate lambda from median survival time}
\usage{
lambda_from_median(median)
}
\arguments{
\item{median}{Median survival time. Must be positive numeric value or vector.}
}
\description{
Hazard rate lambda from median survival time
}
