% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_sum_exp.R
\name{median_sum_exp}
\alias{median_sum_exp}
\title{Median of Sum of Two Exponential Survival Times}
\usage{
median_sum_exp(
  median_pfs,
  median_postpd,
  method = c("formula", "simulation"),
  n_sim = 1e+05,
  seed = NULL,
  upper = 200
)
}
\arguments{
\item{median_pfs}{Median PFS.}

\item{median_postpd}{Median post-PD survival.}

\item{method}{Calculation method, either "formula" or "simulation".}

\item{n_sim}{Number of simulations when method = "simulation".}

\item{seed}{Random seed.}

\item{upper}{Upper bound used in numerical root finding.}
}
\value{
A data.frame containing median total OS and related statistics.
}
\description{
Calls server API to calculate the median of total OS defined as
PFS plus post-PD survival, assuming both are independent exponential distributions.
}
\examples{
\dontrun{
median_sum_exp(
  median_pfs = 3,
  median_postpd = 27,
  method = "formula"
)

median_sum_exp(
  median_pfs = 3,
  median_postpd = 27,
  method = "simulation",
  n_sim = 100000,
  seed = 2024
)
}
}
