% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{oncoseqos_get_api_token}
\alias{oncoseqos_get_api_token}
\title{Get current API token}
\usage{
oncoseqos_get_api_token(mask = TRUE)
}
\arguments{
\item{mask}{Whether to mask the token in display.}
}
\description{
Get current API token
}
