% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{oncoseqos_set_api_token}
\alias{oncoseqos_set_api_token}
\title{Set OncoSeqOS API token}
\usage{
oncoseqos_set_api_token(token)
}
\arguments{
\item{token}{API token provided by the server owner.}
}
\description{
Set OncoSeqOS API token
}
