% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{oncoseqos_set_server}
\alias{oncoseqos_set_server}
\title{Set OncoSeqOS API server}
\usage{
oncoseqos_set_server(url)
}
\arguments{
\item{url}{API server base URL, e.g. "http://116.62.190.134:10000"}
}
\description{
Set OncoSeqOS API server
}
