% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_grid_simulation}
\alias{submit_grid_simulation}
\title{Submit grid simulation job}
\usage{
submit_grid_simulation(
  n_simu = 1000,
  n_total = 282,
  interim_events = 138,
  final_events = 197,
  alpha_interim = 0.0147,
  alpha_final = 0.04551,
  median_pfs_ctl = 3,
  median_pfs_trt = 14.6,
  median_os_ctl_no = 9.5,
  median_postpd_ctl_subseq1 = 22,
  median_postpd_ctl_subseq2 = 15,
  median_os_trt_no = 25,
  median_postpd_trt_subseq1 = 22,
  median_postpd_trt_subseq2 = 15,
  prop_ctl_subseq1 = 0.15,
  prop_ctl_subseq2 = seq(0.1, 0.8, by = 0.05),
  prop_trt_subseq1 = 0.05,
  prop_trt_subseq2 = 0.05,
  hr_thr = 0.8,
  enroll_duration = 11,
  target_censor_rate = 0.29,
  seed = 20260427,
  max_seconds_per_sim = 120
)
}
\arguments{
\item{n_simu}{Integer. Number of simulation replicates.}

\item{n_total}{Integer. Total sample size.}

\item{interim_events}{Integer. Number of events required for interim analysis.}

\item{final_events}{Integer. Number of events required for final analysis.}

\item{alpha_interim}{Numeric. Alpha level for interim analysis.}

\item{alpha_final}{Numeric. Alpha level for final analysis.}

\item{median_pfs_ctl}{Numeric. Median PFS in the control arm.}

\item{median_pfs_trt}{Numeric. Median PFS in the treatment arm.}

\item{median_os_ctl_no}{Numeric. Median OS for control patients without subsequent therapy.}

\item{median_postpd_ctl_subseq1}{Numeric. Median post-progression survival for control patients receiving subsequent therapy 1.}

\item{median_postpd_ctl_subseq2}{Numeric. Median post-progression survival for control patients receiving subsequent therapy 2.}

\item{median_os_trt_no}{Numeric. Median OS for treatment patients without subsequent therapy.}

\item{median_postpd_trt_subseq1}{Numeric. Median post-progression survival for treatment patients receiving subsequent therapy 1.}

\item{median_postpd_trt_subseq2}{Numeric. Median post-progression survival for treatment patients receiving subsequent therapy 2.}

\item{prop_ctl_subseq1}{Numeric. Proportion of control-arm patients receiving subsequent therapy 1.}

\item{prop_ctl_subseq2}{Numeric vector. Proportion values of control-arm patients receiving subsequent therapy 2.}

\item{prop_trt_subseq1}{Numeric. Proportion of treatment-arm patients receiving subsequent therapy 1.}

\item{prop_trt_subseq2}{Numeric. Proportion of treatment-arm patients receiving subsequent therapy 2.}

\item{hr_thr}{Numeric. Hazard ratio threshold.}

\item{enroll_duration}{Numeric. Enrollment duration.}

\item{target_censor_rate}{Numeric. Target censoring rate.}

\item{seed}{Integer. Random seed.}

\item{max_seconds_per_sim}{Integer. Maximum time allowed per simulation (seconds)}
}
\value{
A list containing job submission information, including job ID and status/result URLs.
}
\description{
Submit an asynchronous grid simulation job to the OncoSeqOS API server.
}
